// packages/auth/types/profile.ts
export type Profile = {
  id?: string | number;
  email?: string | null;
  name?: string | null;
  firstName?: string | null;
  fullName?: string | null;
  isAnonymous?: boolean | null;
  emailVerified?: boolean | null;
  image?: string | null;
  avatar?: string | null;
  avatarUrl?: string | null;
};